Use @param and @return in all interface fns (#1265)
* Vault change type of free slot

* Document account module

* Document events

* WIP error review

* WIP errors review

* Finished errors review

* Linter fixes

* Progress documenting interface params

* WIP @params

* Linter fixes

* WIP @params

* WIP @params

* WIP @params

* Linter fixes

* WIP @params

* Minor fixes on params and returns

* Decimal precision in @param and @return

* Linter fixes

* Natspec fixes

* Update protocol/synthetix/contracts/interfaces/IAccountModule.sol


* Update protocol/synthetix/contracts/interfaces/IAccountModule.sol


* Update protocol/synthetix/contracts/interfaces/IAccountModule.sol


* Update protocol/synthetix/contracts/interfaces/ICollateralModule.sol


* Update protocol/synthetix/contracts/interfaces/IUSDTokenModule.sol


* Update protocol/synthetix/contracts/interfaces/IVaultModule.sol


* Update protocol/synthetix/contracts/interfaces/IVaultModule.sol


* Update protocol/synthetix/contracts/interfaces/IVaultModule.sol


* Update protocol/synthetix/contracts/interfaces/IVaultModule.sol


* Reverse param order in AccountCreated event

* accountId first in RewardManagerModule fns

* Linter fixes

* Update protocol/synthetix/contracts/interfaces/ICollateralModule.sol

* Update protocol/synthetix/contracts/interfaces/IIssueUSDModule.sol


* Update protocol/synthetix/contracts/interfaces/IIssueUSDModule.sol


* Update protocol/synthetix/contracts/interfaces/IIssueUSDModule.sol


* Update protocol/synthetix/contracts/interfaces/IUSDTokenModule.sol


* Update protocol/synthetix/contracts/interfaces/IUSDTokenModule.sol


* Update protocol/synthetix/contracts/interfaces/IIssueUSDModule.sol


* Update protocol/synthetix/contracts/interfaces/IRewardsManagerModule.sol


* Update protocol/synthetix/contracts/interfaces/IRewardsManagerModule.sol


* Update protocol/synthetix/contracts/interfaces/IRewardsManagerModule.sol


* Update protocol/synthetix/contracts/interfaces/IRewardsManagerModule.sol


* Update protocol/synthetix/contracts/interfaces/IPoolModule.sol


* Apply suggestions from code review

+++ protocol/synthetix/contracts/interfaces/ICollateralConfigurationModule.sol
@@ -10,11 +10,14 @@ import "../storage/CollateralConfiguration.sol";
 interface ICollateralConfigurationModule {
     /**
      * @notice Emitted when a collateral type’s configuration is created or updated.
+     * @param collateralType The address of the collateral type that was just configured.
+     * @param config The object with the newly configured details.
      */
     event CollateralConfigured(address indexed collateralType, CollateralConfiguration.Data config);
 
     /**
      * @notice Creates or updates the configuration for the given `collateralType`.
+     * @param config The CollateralConfiguration object describing the new configuration.
      *
      * Requirements:
      *
@@ -28,6 +31,8 @@ interface ICollateralConfigurationModule {
     /**
      * @notice Returns a list of detailed information pertaining to all collateral types registered in the system.
      * @dev Optionally returns only those that are currently enabled.
+     * @param hideDisabled Wether to hide disabled collaterals or just return the full list of collaterals in the system.
+     * @return collaterals The list of collateral configuration objects set in the system.
      */
     function getCollateralConfigurations(
         bool hideDisabled
@@ -35,6 +40,8 @@ interface ICollateralConfigurationModule {
 
     /**
      * @notice Returns detailed information pertaining the specified collateral type.
+     * @param collateralType The address for the collateral whose configuration is being queried.
+     * @return collateral The configuration object describing the given collateral.
      */
     function getCollateralConfiguration(
         address collateralType
@@ -42,6 +49,8 @@ interface ICollateralConfigurationModule {
 
     /**
      * @notice Returns the current value of a specified collateral type.
+     * @param collateralType The address for the collateral whose price is being queried.
+     * @return The price of the given collateral, denominated with 18 decimals of precision.
      */
-    function getCollateralPrice(address collateralType) external view returns (uint);
+    function getCollateralPrice(address collateralType) external view returns (uint256);
 }
